refactor: restructure project into logical sub-packages and optimize Docker builds
- Move API entry points to stock_analysis_mcp/api/
- Move core configuration and logging to stock_analysis_mcp/core/
- Move NSE client and stock logic to stock_analysis_mcp/services/
- Update VS Code launch and task configurations to match new structure
- Relocate compose.yml to docker/ and update task paths
- Implement multi-stage Docker builds using uv for optimized image size and speed
- Add .dockerignore and update README.md with new project layout
