+++ .dockerignore
@@ -1,0 +1,22 @@
+.git
+.venv
+.mypy_cache
+.ruff_cache
+.pytest_cache
+.coverage
+htmlcov
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+.env
+.vscode
+.dockerignore
+docker/
+node_modules/
+dist/
+build/
+*.egg-info/
+.DS_Store
+*.swp
+*.bak
