+++ .vscode/launch.json
@@ -5,7 +5,7 @@
             "name": "Python: FastMCP Server",
             "type": "debugpy",
             "request": "launch",
-            "module": "stock_analysis_mcp.server",
+            "module": "stock_analysis_mcp.api.mcp_server",
             "envFile": "${workspaceFolder}/.env",
             "justMyCode": true,
             "console": "integratedTerminal"
@@ -14,7 +14,7 @@
             "name": "Python: ADK Service",
             "type": "debugpy",
             "request": "launch",
-            "module": "stock_analysis_mcp.adk_service",
+            "module": "stock_analysis_mcp.api.adk_server",
             "envFile": "${workspaceFolder}/.env",
             "justMyCode": true,
             "console": "integratedTerminal"
