+++ stock_analysis_mcp/agent/agent.py
@@ -11,7 +11,7 @@ from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset
 from google.adk.tools.tool_context import ToolContext
 
 from stock_analysis_mcp.agent.tools import generate_plot, generate_plot_data_agent
-from stock_analysis_mcp.logging_config import setup_logging
+from stock_analysis_mcp.core.logging_config import setup_logging
 
 setup_logging()
 logger = logging.getLogger(__name__)
