+++ stock_analysis_mcp/agent/tools.py
@@ -13,7 +13,7 @@ from google.adk.tools import ToolContext
 from google.adk.tools.agent_tool import AgentTool
 
 from stock_analysis_mcp.agent.subagent import generate_plot_code_agent
-from stock_analysis_mcp.logging_config import setup_logging
+from stock_analysis_mcp.core.logging_config import setup_logging
 
 setup_logging()
 logger = logging.getLogger(__name__)
