+++ stock_analysis_mcp/api/adk_server.py
@@ -8,7 +8,7 @@ from google.adk.cli.fast_api import get_fast_api_app
 
 load_dotenv()
 
-AGENT_DIR = os.path.dirname(os.path.abspath(__file__))
+AGENT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 app: FastAPI = get_fast_api_app(
     agents_dir=AGENT_DIR,
