+++ stock_analysis_mcp/core/constants.py
